disable wrapping of WebContents by NavigationController
mozbrowserlocationchange doesn't provide the inPage and replaceEntry
metadata that would be required for NavigationController to "compeletely do
history control on user land" (as navigation-controller.js describes it).

So we disable the wrapping of WebContents by NavigationController
and then implement methods that are unique to NavigationController
over in atom_browser_web_contents.js.